Add python 3 compatibility
Mostly, straighten out the use of str/bytes vs unicode/str in ways
that work both in python 2 and in python 3.

In unit testing, add a py3 environment to tox. For
test_install_editable_with_hash_egg_name, use a fork of the repo that
lacks the python 2.7 requirement in its setup.cfg so that python 3 can
work at all.